refactor(auth-hook): change name of hook for authentication
